NotFoundException type fix

+++ src/Clever.Domain/Exceptions/NotFoundException.cs
@@ -4,7 +4,7 @@ namespace Clever.Domain.Exceptions
 {
     public class NotFoundException : Exception
     {
-        public NotFoundException(string entityName, int id) : base($"Entity ({entityName}) with the specified ID ({id}) could not be found") {}
+        public NotFoundException(string entityName, long id) : base($"Entity ({entityName}) with the specified ID ({id}) could not be found") {}
 
     }
 }
